Defer Script Host disposal during cold start
One of the steps during specialization is to restart host. This restart involves starting a new host and disposing the old host.

From profiling, disposal of the old host is the most expensive part in the restart operation. On a single core machine, while the restart is in progress, function invocation will be delayed during cold start. On most cold starts the host dispose seems to be getting scheduled only after cold start has completed so this doesn't impact cold start at lower percentiles. But if the ordering of execution happened to be reversed, the host dispose will increase cold start.

Fix is to defer the Host disposal by 5 seconds. The change reduces P90 cold start by 60%.